{"schema":"decodex/commit/1","summary":"{\"schema\":\"decodex/commit/1\",\"summary\":\"Split live adapter QMD runtime modules\",\"authority\":\"manual\"}","authority":"manual"}
