+++ apps/elf-eval/src/bin/real_world_live_adapter/qmd/checkout.rs
@@ -1,0 +1,27 @@
+use crate::{Command, Path, QmdArgs, Result, fs};
+
+pub(super) fn ensure_qmd_checkout(args: &QmdArgs, log_path: &Path) -> Result<()> {
+	if !args.qmd_dir.exists() {
+		if let Some(parent) = args.qmd_dir.parent() {
+			fs::create_dir_all(parent)?;
+		}
+
+		crate::run_logged_command(
+			"qmd clone",
+			Command::new("git")
+				.arg("clone")
+				.arg("--depth")
+				.arg("1")
+				.arg(&args.qmd_repo_url)
+				.arg(&args.qmd_dir),
+			log_path,
+		)?;
+	}
+
+	crate::run_logged_shell(
+		"qmd install",
+		&args.qmd_dir,
+		"(npm ci || npm install --no-audit --no-fund) && npm run build --if-present",
+		log_path,
+	)
+}
